+++ agent/api/ecsclient/client.go
@@ -109,7 +109,7 @@ func (client *APIECSClient) CreateCluster(clusterName string) (string, error) {
 // instance ARN allows a container instance to update its registered
 // resources.
 func (client *APIECSClient) RegisterContainerInstance(containerInstanceArn string,
-	attributes []*ecs.Attribute, tags []*ecs.Tag, platformDevices []*ecs.PlatformDevice) (string, string, error) {
+	attributes []*ecs.Attribute, tags []*ecs.Tag, registrationToken string, platformDevices []*ecs.PlatformDevice) (string, string, error) {
 	clusterRef := client.config.Cluster
 	// If our clusterRef is empty, we should try to create the default
 	if clusterRef == "" {
@@ -120,7 +120,7 @@ func (client *APIECSClient) RegisterContainerInstance(containerInstanceArn strin
 		}()
 		// Attempt to register without checking existence of the cluster so we don't require
 		// excess permissions in the case where the cluster already exists and is active
-		containerInstanceArn, availabilityzone, err := client.registerContainerInstance(clusterRef, containerInstanceArn, attributes, tags, platformDevices)
+		containerInstanceArn, availabilityzone, err := client.registerContainerInstance(clusterRef, containerInstanceArn, attributes, tags, registrationToken, platformDevices)
 		if err == nil {
 			return containerInstanceArn, availabilityzone, nil
 		}
@@ -131,11 +131,11 @@ func (client *APIECSClient) RegisterContainerInstance(containerInstanceArn strin
 			return "", "", err
 		}
 	}
-	return client.registerContainerInstance(clusterRef, containerInstanceArn, attributes, tags, platformDevices)
+	return client.registerContainerInstance(clusterRef, containerInstanceArn, attributes, tags, registrationToken, platformDevices)
 }
 
 func (client *APIECSClient) registerContainerInstance(clusterRef string, containerInstanceArn string,
-	attributes []*ecs.Attribute, tags []*ecs.Tag, platformDevices []*ecs.PlatformDevice) (string, string, error) {
+	attributes []*ecs.Attribute, tags []*ecs.Tag, registrationToken string, platformDevices []*ecs.PlatformDevice) (string, string, error) {
 	registerRequest := ecs.RegisterContainerInstanceInput{Cluster: &clusterRef}
 	var registrationAttributes []*ecs.Attribute
 	if containerInstanceArn != "" {
@@ -170,6 +170,8 @@ func (client *APIECSClient) registerContainerInstance(clusterRef string, contain
 	}
 
 	registerRequest.TotalResources = resources
+
+	registerRequest.ClientToken = &registrationToken
 	resp, err := client.standardClient.RegisterContainerInstance(&registerRequest)
 	if err != nil {
 		seelog.Errorf("Unable to register as a container instance with ECS: %v", err)