+++ agent/app/agent_capability.go
@@ -38,6 +38,7 @@ const (
 	attributeSeparator                          = "."
 	capabilityPrivateRegistryAuthASM            = "private-registry-authentication.secretsmanager"
 	capabilitySecretEnvSSM                      = "secrets.ssm.environment-variables"
+	capabilitySecretEnvASM                      = "secrets.asm.environment-variables"
 	capabiltyPIDAndIPCNamespaceSharing          = "pid-ipc-namespace-sharing"
 	capabilityNvidiaDriverVersionInfix          = "nvidia-driver-version."
 	capabilityECREndpoint                       = "ecr-endpoint"
@@ -72,6 +73,7 @@ const (
 //    ecs.capability.secrets.ssm.environment-variables
 //    ecs.capability.pid-ipc-namespace-sharing
 //    ecs.capability.ecr-endpoint
+//    ecs.capability.secrets.asm.environment-variables
 
 func (agent *ecsAgent) capabilities() ([]*ecs.Attribute, error) {
 	var capabilities []*ecs.Attribute
@@ -131,6 +133,9 @@ func (agent *ecsAgent) capabilities() ([]*ecs.Attribute, error) {
 	}
 	// support ecr endpoint override
 	capabilities = appendNameOnlyAttribute(capabilities, attributePrefix+capabilityECREndpoint)
+
+	// ecs agent version 1.23.0 supports ecs secrets integrating with aws secrets manager
+	capabilities = appendNameOnlyAttribute(capabilities, attributePrefix+capabilitySecretEnvASM)
 
 	return capabilities, nil
 }