+++ agent/taskresource/types/types.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
 	asmauthres "github.com/aws/amazon-ecs-agent/agent/taskresource/asmauth"
+	asmsecretres "github.com/aws/amazon-ecs-agent/agent/taskresource/asmsecret"
 	cgroupres "github.com/aws/amazon-ecs-agent/agent/taskresource/cgroup"
 	ssmsecretres "github.com/aws/amazon-ecs-agent/agent/taskresource/ssmsecret"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/volume"
@@ -33,6 +34,8 @@ const (
 	ASMAuthKey = asmauthres.ResourceName
 	// SSMSecretKey is the string used in resources map to represent ssm secret
 	SSMSecretKey = ssmsecretres.ResourceName
+	// ASMSecretKey is the string used in resources map to represent asm secret
+	ASMSecretKey = asmsecretres.ResourceName
 )
 
 // ResourcesMap represents the map of resource type to the corresponding resource
@@ -63,6 +66,10 @@ func (rm *ResourcesMap) UnmarshalJSON(data []byte) error {
 			}
 		case SSMSecretKey:
 			if unmarshalSSMSecretKey(key, value, result) != nil {
+				return err
+			}
+		case ASMSecretKey:
+			if unmarshalASMSecretKey(key, value, result) != nil {
 				return err
 			}
 		default:
@@ -134,6 +141,24 @@ func unmarshalSSMSecretKey(key string, value json.RawMessage, result map[string]
 
 	for _, secret := range ssmsecrets {
 		res := &ssmsecretres.SSMSecretResource{}
+		err := res.UnmarshalJSON(secret)
+		if err != nil {
+			return err
+		}
+		result[key] = append(result[key], res)
+	}
+	return nil
+}
+
+func unmarshalASMSecretKey(key string, value json.RawMessage, result map[string][]taskresource.TaskResource) error {
+	var asmsecrets []json.RawMessage
+	err := json.Unmarshal(value, &asmsecrets)
+	if err != nil {
+		return err
+	}
+
+	for _, secret := range asmsecrets {
+		res := &asmsecretres.ASMSecretResource{}
 		err := res.UnmarshalJSON(secret)
 		if err != nil {
 			return err